moving raii wrapper from qnn execution provider
